+++ src/main/java/com/ddook/ddak/board/repository/ArticlesRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ddook.ddak.board.model.Article;
 
-public interface ArticlesRepository  extends JpaRepository<Article, Long> {
+public interface ArticlesRepository  extends JpaRepository<Article, Integer> {
 }
